requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Gemini AI
google-generativeai==0.3.2

# Data validation
pydantic==2.5.3

# Database
sqlalchemy==2.0.25

# Environment
python-dotenv==1.0.0

# WebSocket
websockets==12.0

# PDF Generation
reportlab==4.0.8

# Async support
aiofiles==23.2.1

# CORS
starlette==0.35.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
